python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ridesecure-ml"
version = "1.0.0"
description = "Machine learning components for RideSecure helmet detection system"
authors = [{name = "RideSecure Team"}]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]

dependencies = [
    "torch>=1.12.0",
    "torchvision>=0.13.0",
    "ultralytics>=8.0.0",
    "opencv-python>=4.6.0",
    "easyocr>=1.7.0",
    "numpy>=1.21.0",
    "pillow>=8.3.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "scikit-learn>=1.1.0",
    "pandas>=1.4.0",
    "tqdm>=4.64.0",
    "pyyaml>=6.0",
    "onnx>=1.12.0",
    "onnxruntime>=1.12.0",
    "flask>=2.2.0",
    "fastapi>=0.85.0",
    "uvicorn>=0.18.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.1.0",
    "pytest-cov>=3.0.0",
    "black>=22.3.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
    "pre-commit>=2.19.0",
]

[project.urls]
Homepage = "https://github.com/singharyan006/ride-secure"
Repository = "https://github.com/singharyan006/ride-secure.git"
Issues = "https://github.com/singharyan006/ride-secure/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=ridesecure --cov-report=term-missing"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true